getting rid of migration dir, moving person and household outwards
