+++ src/vivarium_census_prl_synth_pop/components/__init__.py
@@ -1,8 +1,8 @@
 from vivarium_census_prl_synth_pop.components.population import Population
 from vivarium_census_prl_synth_pop.components.observers import Observers
 from vivarium_census_prl_synth_pop.components.synthetic_pii import Address
-from vivarium_census_prl_synth_pop.components.migration.household import HouseholdMigration
-from vivarium_census_prl_synth_pop.components.migration.person import PersonMigration
+from vivarium_census_prl_synth_pop.components.household import HouseholdMigration
+from vivarium_census_prl_synth_pop.components.person import PersonMigration
 from vivarium_census_prl_synth_pop.components.mortality import Mortality
 from vivarium_census_prl_synth_pop.components.fertility import Fertility
 from vivarium_census_prl_synth_pop.components.businesses import Businesses
